update positional arguments and fixes
